+++ src/arp_test/ARPDlg.java
@@ -1,4 +1,4 @@
-package stopwait;
+package arp_test;
 
 import java.awt.Color;
 import java.awt.Container;
