+++ src/arp_test/EthernetLayer.java
@@ -1,4 +1,4 @@
-package stopwait;
+package arp_test;
 
 import java.util.ArrayList;
 import java.util.StringTokenizer;
